mujoco_sim/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "halo-mujoco-sim"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "h5py>=3.0",
    "tqdm>=4.60",
]

[project.optional-dependencies]
sim = ["mujoco>=3.1.6"]
viewer = ["opencv-python>=4.8"]
server = ["pyzmq>=25", "msgpack>=1.0", "opencv-python>=4.8"]

[tool.hatch.build.targets.wheel]
packages = ["mujoco_sim"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
